Dialog for confirm delete/save note.

+++ NoteMe/NoteMe/ViewModels/NoteEntryPageViewModel.cs
@@ -2,6 +2,7 @@
 using NoteMe.Services;
 using Prism.Commands;
 using Prism.Navigation;
+using Prism.Services;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -12,6 +13,7 @@ namespace NoteMe.ViewModels
     public class NoteEntryPageViewModel : ViewModelBase, INavigationAware
     {
         private readonly INoteService _noteService;
+        private readonly IPageDialogService _dialogService;
         public Note _note;
         public Note NoteToEdit
         {
@@ -27,23 +29,46 @@ namespace NoteMe.ViewModels
         public DelegateCommand SaveNoteCommand { get; set; }
         public DelegateCommand DeleteNoteCommand { get; set; }
 
-        public NoteEntryPageViewModel(INavigationService navigationService, INoteService noteService) : base(navigationService)
+        public NoteEntryPageViewModel(INavigationService navigationService, INoteService noteService, IPageDialogService dialogService) : base(navigationService)
         {
             NoteToEdit = new Note();
             SaveNoteCommand = new DelegateCommand(SaveNote);
             DeleteNoteCommand = new DelegateCommand(DeleteNoteAsync);
             _noteService = noteService;
+            _dialogService = dialogService;
         }
 
         private async void DeleteNoteAsync()
         {
-            await _noteService.DeleteNoteAsync(NoteToEdit);
-            await NavigationService.GoBackAsync();
+            bool delete = await _dialogService.DisplayAlertAsync("Alert", "Delete this note?", "Continue", "Cancel");
+
+            if (delete)
+            {
+                int result = await _noteService.DeleteNoteAsync(NoteToEdit);
+
+                if(result > 0)
+                {
+                    await _dialogService.DisplayAlertAsync("Alert", "Your note has been deleted!!", "OK");
+                }
+                else
+                {
+                    await _dialogService.DisplayAlertAsync("Alert", "Something went wrong. The note wasn't deleted!!", "OK");
+                }
+                await NavigationService.GoBackAsync();
+            }
         }
 
         private async void SaveNote()
         {
-            await _noteService.SaveNoteAsync(NoteToEdit);
+            int result = await _noteService.SaveNoteAsync(NoteToEdit);
+            if (result > 0)
+            {
+                await _dialogService.DisplayAlertAsync("Alert", "Your note has been saved!!", "OK");
+            }
+            else
+            {
+                await _dialogService.DisplayAlertAsync("Alert", "Something went wrong. The note wasn't saved!!", "OK");
+            }
             await NavigationService.GoBackAsync();
         }
